refactor(core): hoist non-vite modules out of vite/ into editing/, files/, http/
After the comments/files-plugin splits, 60% of src/vite/ was pure logic
with no Vite dependency, just consumed by Vite plugins. Move those out
so the directory name matches what's actually inside:

- editing/ — comments, edit-ops, revert-asset, slide-ops, babel-walk
- files/   — assets, folders
- http/    — request-guard
- vite/    — only Vite plugins + config
