+++ packages/core/src/files/assets.test.ts
