+++ packages/core/src/files/folders.test.ts
